713 moved Activity2 into Activity
